Started to test how the api works
Had to change name of Amazon class to Agent because Amazon is a class in Gao's package.

So far i created a new player from the Agent class that is auto joining okanagan-room. I have the handle message somewhat working and i've noticed that the
Agent's GAME_ACTION_MOVE case logs the what human gui player is doing.

I still need to figure out the following:
How to tell if youre white or black,.
how to decode the board values returned in GAME_STATE_BOARD.
How to update the game board with moves